fix(approval): resolve compilation errors regarding fileSizeKb and CommonCode.name

+++ backend/src/main/java/com/tphr/hr/approval/dto/ApprovalDto.java
@@ -48,7 +48,7 @@ public class ApprovalDto {
         private Long id;
         private String fileName;
         private String filePath;
-        private Long fileSizeKb;
+        private Integer fileSizeKb;
     }
 
     @Getter
